APU and CPU: Bug fixes
Frame counter quarter and half cycles are now executed when mode flag is set
when writing to $4017. Frame counter state is also nwow reset properly.

A bug with CPU interrupts was also fixed, where the emulation would get
stuck in an infinite loop if interrupt flag is set when CPU ignores
interrupts.

Also adds tons of tests and accuracy fixes for APU, based on test roms